+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The scope of what is covered by the version number excludes:
 
 - Feat: when detecting character display width, also accept unicode codepoints (integers),
   since the Lua utf8 library returns codepoints, not strings
+- Feat: allow passing in a sleep function to `readkey` and `readansi`
 - Fix: NetBSD fix compilation, undeclared directives
 - Refactor: random bytes; remove deprecated API usage on Windows, move to
   binary api instead of /dev/urandom file on linux and bsd